Restructure payment and ticket management actions
