+++ src/actions/payments.actions.ts
@@ -1,13 +1,10 @@
 'use server';
 
-import { OrderPaymentStatus, Payment, PaymentSite, PaymentStatus, PaymentMethod, Prisma, OrderStatus, User } from "@prisma/client";
+import { Payment, User } from "@prisma/client";
 import prisma from "@/lib/db";
 import { QueryParams, PaginatedResponse } from "@/types/common";
 import { GetObjectByTParams } from "@/types/extended";
-import { createLog } from '@/actions/logs.actions';
-import { sendOrderStatusEmail, sendPaymentStatusEmail } from "@/lib/email-service";
 import { calculatePagination, processActionReturnData, verifyPermission } from "@/utils";
-import serverSideEffect from "@/utils/serverSideEffect";
 
 
 /**
@@ -302,787 +299,4 @@ export async function getPaymentsByOrderId({ userId, orderId, limitFields }: Get
     success: true,
     data: processActionReturnData(payments, limitFields) as Payment[]
   };
-}
-
-/**
- * Processes a payment for a specified order.
- *
- * This function verifies that the user has permission to process payments, retrieves the associated order from
- * the database (including its payments and customer details), and calculates the new total paid by summing only the
- * payments with a 'VERIFIED' status along with the new payment amount. If the order exists and the total paid does
- * not exceed the order's total amount, a new payment record is created. The order's payment status is then updated
- * (and, if the order becomes fully paid while initially pending, its status is changed to 'PROCESSING'). The function
- * logs the operation, attempts to send a payment status notification email with relevant order and customer details
- * (logging any email errors without affecting the payment process), and invalidates related caches.
- *
- * @param userId - The ID of the user processing the payment.
- * @param orderId - The ID of the order to which the payment is applied.
- * @param amount - The payment amount.
- * @param paymentMethod - The method used for the payment.
- * @param paymentSite - The site or point from which the payment is processed.
- * @param paymentStatus - The status of the payment (e.g., pending, completed).
- * @param referenceNo - (Optional) A reference number for the payment.
- * @param memo - (Optional) A memo or note for the payment.
- * @param transactionId - (Optional) A unique transaction identifier.
- * @param paymentProvider - (Optional) The provider handling the payment.
- * @param limitFields - (Optional) An array of field names to remove from the returned payment object.
- *
- * @returns A Promise resolving to an object containing:
- *  - success: A boolean indicating whether the payment processing was successful.
- *  - data: The processed Payment object with sensitive fields removed (if `limitFields` is provided) on success.
- *  - message: An error message if the operation fails.
- *  - errors: (Optional) Additional error details.
- *
- * @example
- * const result = await processPayment({
- *   userId: "user123",
- *   orderId: "order456",
- *   amount: 1000,
- *   paymentMethod: PaymentMethod.CREDIT_CARD,
- *   paymentSite: PaymentSite.ONLINE,
- *   paymentStatus: PaymentStatus.COMPLETED,
- *   transactionId: "txn789"
- * });
- *
- * if (result.success) {
- *   console.log("Payment processed:", result.data);
- * } else {
- *   console.error("Payment failed:", result.message);
- * }
- */
-export async function processPayment({ 
-  userId, 
-  orderId,
-  amount,
-  paymentMethod,
-  paymentSite,
-  paymentStatus,  
-  referenceNo,
-  memo,
-  transactionId,
-  paymentProvider,
-  limitFields
-}: {
-  userId: string;
-  orderId: string;
-  amount: number;
-  paymentMethod: PaymentMethod;
-  paymentSite: PaymentSite;
-  paymentStatus: PaymentStatus;
-  referenceNo?: string;
-  memo?: string;
-  transactionId?: string;
-  paymentProvider?: string;
-  limitFields?: string[];
-}): Promise<ActionsReturnType<Payment>> {
-  try {
-    const order = await prisma.order.findUnique({
-      where: { id: orderId },
-      include: {
-        payments: true,
-        customer: true
-      }
-    });
-    if (!order) {
-      await createLog({
-        userId: userId,
-        createdById: userId,
-        reason: "Payment Processing Failed - Invalid Order",
-        systemText: `User attempted to process payment for non-existent order ${orderId}`,
-        userText: "Order not found."
-      });
-      return {
-        success: false,
-        message: "Order not found."
-      };
-    }
-    const totalPaid = order.payments.filter(payment => payment.paymentStatus === 'VERIFIED').reduce((sum, payment) => sum + Number(payment.amount), 0) + amount;
-    if (totalPaid > Number(order.totalAmount)) {
-      await createLog({
-        userId: order.customerId,
-        createdById: userId,
-        reason: "Payment Processing Failed - Amount Exceeds Total",
-        systemText: `Payment amount ${amount} would exceed order total ${order.totalAmount} for order ${orderId}`,
-        userText: "Payment amount exceeds order total."
-      });
-      return {
-        success: false,
-        message: "Payment amount exceeds order total."
-      };
-    }
-    const payment = await prisma.payment.create({
-      data: {
-        orderId,
-        userId: order.customerId,
-        processedById: userId,
-        amount: new Prisma.Decimal(amount),
-        paymentMethod,
-        paymentSite,
-        paymentStatus,
-        referenceNo: referenceNo ?? "",
-        memo,
-        transactionId,
-        paymentProvider
-      },
-      include: {
-        order: true,
-        user: true
-      }
-    });
-
-    // For offsite pending payments, don't update order status or send emails
-    const isOffsitePending = paymentSite === PaymentSite.OFFSITE && paymentStatus === PaymentStatus.PENDING;
-    
-    if (!isOffsitePending) {
-      // Only update order status for non-offsite or verified payments
-      const isFullyPaid = totalPaid === Number(order.totalAmount);
-      
-      await prisma.order.update({
-        where: { id: orderId },
-        data: { 
-          paymentStatus: isFullyPaid ? OrderPaymentStatus.PAID : OrderPaymentStatus.DOWNPAYMENT,
-          // Change order status to PROCESSING only if fully paid and current status is PENDING
-          ...(isFullyPaid && order.status === 'PENDING' ? { status: 'PROCESSING' } : {})
-        }
-      });
-
-      // Create success log with appropriate messaging
-      await createLog({
-        userId: order.customerId,
-        createdById: userId,
-        reason: "Payment Processed Successfully",
-        systemText: `Payment of ₱${amount} processed for order ${orderId} via ${paymentMethod}. Status: ${paymentStatus}. ${isFullyPaid ? 'Order is now fully paid.' : ''}`,
-        userText: `Payment of ₱${amount} has been processed successfully${isFullyPaid ? '. Your order is now fully paid.' : '.'}`
-      });
-
-      // Send payment notification email for verified payments only
-      try {
-        serverSideEffect(
-          () => sendPaymentStatusEmail({
-            orderNumber: orderId,
-            customerName: `${order.customer.firstName} ${order.customer.lastName}`,
-            customerEmail: order.customer.email,
-            amount,
-            status: 'verified'
-          })
-        );
-      } catch (emailError) {
-        serverSideEffect(
-          () => createLog({
-            userId: order.customerId,
-            createdById: userId,
-            reason: "Payment Notification Email Error",
-            systemText: `Error sending payment notification email for order ${orderId}: ${emailError instanceof Error ? emailError.message : 'Unknown error'}`,
-            userText: "An error occurred while sending the payment notification email."
-          })
-        );
-      }
-    } else {
-      serverSideEffect(
-        () => sendPaymentStatusEmail({
-          orderNumber: orderId,
-          customerName: `${order.customer.firstName} ${order.customer.lastName}`,
-          customerEmail: order.customer.email,
-          amount,
-          status: 'submitted',
-        }),
-        () => createLog({
-          userId: order.customerId,
-          createdById: userId,
-          reason: "Offsite Payment Submitted",
-          systemText: `Offsite payment of ₱${amount} submitted for order ${orderId} via ${paymentMethod}. Awaiting verification.`,
-          userText: `Your payment submission of ₱${amount} has been received and is awaiting verification by our team.`
-        })
-      );
-    }
-    
-    return {
-      success: true,
-      data: processActionReturnData(payment, limitFields) as Payment
-    };
-  } catch (error) {
-    serverSideEffect(
-      () => createLog({
-        userId: userId,
-        createdById: userId,
-        reason: "Payment Processing Error",
-        systemText: `Error processing payment for order ${orderId}: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        userText: "An error occurred while processing the payment."
-      })
-    );
-    
-    return {
-      success: false,
-      message: "Error processing payment.",
-      errors: { error }
-    };
-  }
-}
-
-/**
- * Refunds a payment and updates the corresponding order's payment status.
- *
- * This asynchronous function processes a payment refund by first verifying the user's permissions. It retrieves the payment
- * record from the database and ensures that the refund amount does not exceed the original payment amount. If the conditions are met,
- * a refund record is created with a negative amount to indicate the refund. The function then aggregates the remaining payments for the order
- * to determine and update the order's payment status (e.g., marking the order as fully refunded if no payments remain, or as a downpayment otherwise).
- * Additionally, the function logs key events (such as unauthorized access, non-existent payments, invalid amounts, and errors), sends a refund
- * notification email with order and customer details, and invalidates related caches to ensure data integrity.
- *
- * @param userId - The ID of the user initiating the refund.
- * @param paymentId - The unique identifier of the payment to refund.
- * @param amount - The refund amount, which must not exceed the original payment amount.
- * @param reason - The reason for initiating the refund.
- *
- * @returns A Promise that resolves to an object indicating the result of the refund operation. On success, the object contains a `data` property
- * with the newly created refund payment record; on failure, it includes a `message` explaining the error and may also include an `errors` object with additional details.
- */
-export async function refundPayment(
-  userId: string,
-  paymentId: string,
-  amount: number,
-  reason: string
-): Promise<ActionsReturnType<Payment>> {
-  if (!await verifyPermission({
-    userId,
-    permissions: {
-      payments: { canRead: true, canUpdate: true },
-    }
-  })) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Refund Failed - Unauthorized",
-        systemText: `Unauthorized attempt to refund payment ${paymentId}`,
-        userText: "You are not authorized to process refunds."
-      })
-    );
-    return {
-      success: false,
-      message: "You are not authorized to process refunds."
-    };
-  }
-
-  try {
-    const payment = await prisma.payment.findUnique({
-      where: { id: paymentId },
-      include: {
-        order: true,
-        user: true
-      }
-    });
-
-    if (!payment) {
-      serverSideEffect(
-        () => createLog({
-          userId,
-          createdById: userId,
-          reason: "Payment Refund Failed - Not Found",
-          systemText: `Attempted to refund non-existent payment ${paymentId}`,
-          userText: "Payment not found."
-        })
-      );
-      return {
-        success: false,
-        message: "Payment not found"
-      };
-    }
-
-    // Check if amount to refund is valid
-    if (amount > Number(payment.amount)) {
-      serverSideEffect(
-        () => createLog({
-          userId: payment.userId,
-          createdById: userId,
-          reason: "Payment Refund Failed - Invalid Amount",
-          systemText: `Refund amount ${amount} exceeds original payment amount ${payment.amount} for payment ${paymentId}`,
-          userText: "Refund amount cannot exceed the original payment amount."
-        })
-      );
-      return {
-        success: false,
-        message: "Refund amount cannot exceed the original payment amount"
-      };
-    }
-
-    // Create a refund record
-    const refund = await prisma.payment.create({
-      data: {
-        orderId: payment.orderId,
-        userId: payment.userId,
-        processedById: userId,
-        amount: new Prisma.Decimal(-amount), // Negative amount to indicate refund
-        paymentMethod: payment.paymentMethod,
-        paymentSite: payment.paymentSite,
-        paymentStatus: PaymentStatus.REFUNDED,
-        referenceNo: payment.referenceNo,
-        memo: `Refund for payment ${payment.id}. Reason: ${reason}`,
-        transactionId: payment.transactionId,
-        paymentProvider: payment.paymentProvider
-      },
-      include: {
-        order: {
-          include: {
-            customer: {
-              select: { 
-                email: true,
-                firstName: true,
-                lastName: true
-              }
-            }
-          }
-        },
-        user: true
-      }
-    });
-
-    // Update order payment status if full refund
-    const totalPaidAfterRefund = await prisma.payment.aggregate({
-      where: { orderId: payment.orderId },
-      _sum: { amount: true }
-    });
-
-    const newOrderPaymentStatus = Number(totalPaidAfterRefund._sum.amount) <= 0 
-      ? OrderPaymentStatus.REFUNDED 
-      : OrderPaymentStatus.DOWNPAYMENT;
-
-    await prisma.order.update({
-      where: { id: payment.orderId },
-      data: { 
-        paymentStatus: newOrderPaymentStatus
-      }
-    });
-
-    serverSideEffect(
-      () => createLog({
-        userId: payment.userId,
-        createdById: userId,
-        reason: "Payment Refunded",
-        systemText: `Refund of ₱${amount} processed for payment ${paymentId}. Reason: ${reason}`,
-        userText: `Refund of ₱${amount} has been processed successfully. Reason: ${reason}`
-      }),
-      () => sendPaymentStatusEmail({
-        orderNumber: refund.order.id,
-        customerName: `${refund.order.customer.firstName} ${refund.order.customer.lastName}`,
-        customerEmail: refund.order.customer.email,
-        amount: Number(refund.amount),
-        status: 'refunded',
-        refundReason: reason
-      })
-    );
-
-    return {
-      success: true,
-      data: processActionReturnData(refund) as Payment
-    };
-  } catch (error) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Refund Error",
-        systemText: `Error refunding payment ${paymentId}: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        userText: "An error occurred while processing the refund."
-      })
-    );
-    
-    return {
-      success: false,
-      message: "Error processing refund.",
-      errors: { error }
-    };
-  }
-}
-
-/**
- * Validates a payment for an order by verifying user permissions, checking for duplicate transactions, updating the payment record, updating the order status, sending a verification email, and invalidating caches.
- *
- * This function performs the following steps:
- * - Verifies that the user has the necessary dashboard read permissions.
- * - Retrieves the order and payment by their identifiers and ensures they exist.
- * - Checks for an existing verified payment with the same transaction ID to prevent duplicate processing.
- * - Updates the existing payment record with VERIFIED status in the database.
- * - Calculates the total paid amount and updates the order's payment status to either PAID (if fully paid) or DOWNPAYMENT.
- * - Logs the outcome for auditing purposes.
- * - Sends a payment verification email to the customer with order and payment details.
- * - Invalidates related cache entries.
- *
- * @param userId - The ID of the user performing the payment validation.
- * @param orderId - The identifier of the order for which the payment is being validated.
- * @param amount - The payment amount to be validated.
- * @param transactionDetails - An object containing payment transaction details:
- *   - transactionId: The unique transaction identifier.
- *   - referenceNo: The payment reference number.
- *   - paymentMethod: The method used for the payment.
- *   - paymentSite: The site where the payment was made.
- * @param paymentId - The identifier of the payment to be validated.
- *
- * @returns A promise that resolves to an ActionsReturnType object. On success, the object indicates success; on failure, it contains an error message.
- */
-export async function validatePayment(
-  userId: string,
-  orderId: string,
-  amount: number,
-  transactionDetails: {
-    transactionId: string;
-    referenceNo: string;
-    paymentMethod: PaymentMethod;
-    paymentSite: PaymentSite;
-  },
-  paymentId: string
-): Promise<ActionsReturnType<void>> {
-  if (!await verifyPermission({
-    userId,
-    permissions: {
-      payments: { canRead: true, canUpdate: true },
-    }
-  })) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Validation Failed - Unauthorized",
-        systemText: `Unauthorized attempt to validate payment for order ${orderId}`,
-        userText: "You are not authorized to validate payments."
-      })
-    );
-    return {
-      success: false,
-      message: "You are not authorized to validate payments."
-    };
-  }
-
-  try {
-    // Find the order and validate it exists
-    const order = await prisma.order.findUnique({
-      where: { id: orderId },
-      include: {
-        payments: true,
-        customer: true
-      }
-    });
-
-    if (!order) {
-      serverSideEffect(
-        () => createLog({
-          userId,
-          createdById: userId,
-          reason: "Payment Validation Failed - Order Not Found",
-          systemText: `Attempted to validate payment for non-existent order ${orderId}`,
-          userText: "Order not found."
-        })
-      );
-      return {
-        success: false,
-        message: "Order not found."
-      };
-    }
-
-    // Find the payment to validate
-    const existingPayment = await prisma.payment.findUnique({
-      where: { id: paymentId }
-    });
-
-    if (!existingPayment) {
-      serverSideEffect(
-        () => createLog({
-          userId,
-          createdById: userId,
-          reason: "Payment Validation Failed - Payment Not Found",
-          systemText: `Attempted to validate non-existent payment ${paymentId}`,
-          userText: "Payment not found."
-        })
-      );
-      return {
-        success: false,
-        message: "Payment not found."
-      };
-    }
-
-    // // Check for duplicate verified transaction (with same transactionId)
-    // const duplicateTransaction = await prisma.payment.findFirst({
-    //   where: {
-    //     transactionId: transactionDetails.transactionId,
-    //     paymentStatus: PaymentStatus.VERIFIED,
-    //     id: { not: paymentId } // Exclude the current payment we're validating
-    //   }
-    // });
-
-    // if (duplicateTransaction) {
-    //   await createLog({
-    //     userId: order.customerId,
-    //     createdById: userId,
-    //     reason: "Payment Validation Failed - Duplicate Transaction",
-    //     systemText: `Duplicate transaction ID ${transactionDetails.transactionId} detected for order ${orderId}`,
-    //     userText: "This transaction has already been processed."
-    //   });
-    //   return {
-    //     success: false,
-    //     message: "This transaction has already been processed."
-    //   };
-    // }
-
-    const validatingEmployee = await prisma.user.findUnique({
-      where: { id: userId }
-    });
-    const employeeName = validatingEmployee ? `${validatingEmployee.firstName} ${validatingEmployee.lastName}` : userId;
-
-    // Update the existing payment record
-    const verifiedPayment = await prisma.payment.update({
-      where: { id: paymentId },
-      data: {
-        processedById: userId,
-        paymentStatus: PaymentStatus.VERIFIED,
-        transactionId: transactionDetails.transactionId,
-        referenceNo: transactionDetails.referenceNo,
-        memo: existingPayment.memo ? `${existingPayment.memo} | Payment validated by ${employeeName}` : `Payment validated by ${employeeName}`
-      },
-      include: {
-        order: {
-          include: {
-            customer: {
-              select: {
-                email: true,
-                firstName: true,
-                lastName: true,
-              }
-            }
-          }
-        }
-      }
-    });
-
-    // Update order status if needed
-    const totalPaid = order.payments.reduce((sum, payment) => {
-      if (payment.id === paymentId) {
-        // Skip the payment we're validating as we'll add it separately
-        return sum;
-      }
-      return payment.paymentStatus === 'VERIFIED' ? sum + Number(payment.amount) : sum;
-    }, 0) + Number(existingPayment.amount);
-
-    const isFullyPaid = totalPaid >= Number(order.totalAmount);
-
-    if (isFullyPaid) {
-      const fullypaidOrder = await prisma.order.update({
-        where: { id: orderId },
-        data: {
-          paymentStatus: OrderPaymentStatus.PAID,
-          status: order.status === OrderStatus.PENDING ? OrderStatus.PROCESSING : order.status
-        },
-        include: {
-          customer: true,
-          orderItems: {
-            include: {
-              variant: {
-                include: {
-                  product: true,
-                }
-              }
-            }
-          }
-        }
-      });
-
-      serverSideEffect(
-        () => sendOrderStatusEmail({
-          customerEmail: fullypaidOrder.customer.email,
-          customerName: `${fullypaidOrder.customer.firstName} ${fullypaidOrder.customer.lastName}`,
-          orderNumber: fullypaidOrder.id,
-          newStatus: fullypaidOrder.status,
-          // @ts-expect-error - data is enough already
-          order: fullypaidOrder
-        }),
-      );
-
-    } else {
-      await prisma.order.update({
-        where: { id: orderId },
-        data: {
-          paymentStatus: OrderPaymentStatus.DOWNPAYMENT
-        }
-      });
-    }
-
-    serverSideEffect(
-      () => createLog({
-        userId: order.customerId,
-        createdById: userId,
-        reason: "Payment Validated Successfully",
-        systemText: `Validated payment of ₱${existingPayment.amount} for order ${orderId}. Payment ID: ${paymentId}. Transaction ID: ${transactionDetails.transactionId}`,
-        userText: "Payment has been validated successfully."
-      }),
-      () => sendPaymentStatusEmail({
-        orderNumber: verifiedPayment.order.id,
-        customerName: `${verifiedPayment.order.customer.firstName} ${verifiedPayment.order.customer.lastName}`,
-        customerEmail: verifiedPayment.order.customer.email,
-        amount: Number(verifiedPayment.amount),
-        status: 'verified',
-      })
-    );
-
-    return {
-      success: true
-    };
-  } catch (error) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Validation Error",
-        systemText: `Error validating payment for order ${orderId}: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        userText: "An error occurred while validating the payment."
-      })
-    );
-
-    return {
-      success: false,
-      message: error instanceof Error ? error.message : "Failed to validate payment"
-    };
-  }
-}
-
-/**
- * Rejects a payment for a specified order.
- *
- * This function first verifies that the user has the necessary permissions to reject payments. It then attempts
- * to find the associated order and payment. If both exist, it updates the existing payment record with a "DECLINED" status
- * and adds a memo containing the rejection reason. The function logs the rejection action, invalidates cache entries
- * related to the order and its payments, and returns a success status. In case of permission failure, order or payment not found,
- * or any other error, it logs the error and returns a failure status with an appropriate message.
- *
- * @param userId - ID of the user initiating the rejection.
- * @param orderId - ID of the order associated with the payment.
- * @param paymentId - ID of the payment to be rejected.
- * @param rejectionReason - Explanation for why the payment is being rejected.
- *
- * @returns A promise resolving to an object indicating the success status and an optional error message.
- *
- * @example
- * const result = await rejectPayment("user123", "order456", "payment789", "Payment method declined by bank");
- * if (!result.success) {
- *   console.error(result.message);
- * }
- */
-export async function rejectPayment(
-  userId: string,
-  orderId: string,
-  paymentId: string,
-  rejectionReason: string
-): Promise<ActionsReturnType<void>> {
-  if (!await verifyPermission({
-    userId,
-    permissions: {
-      payments: { canRead: true, canUpdate: true },
-    }
-  })) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Rejection Failed - Unauthorized",
-        systemText: `Unauthorized attempt to reject payment for order ${orderId}`,
-        userText: "You are not authorized to reject payments."
-      })
-    );
-    return {
-      success: false,
-      message: "You are not authorized to reject payments."
-    };
-  }
-
-  try {
-    const order = await prisma.order.findUnique({
-      where: { id: orderId },
-      include: {
-        customer: true
-      }
-    });
-
-    if (!order) {
-      serverSideEffect(
-        () => createLog({
-          userId,
-          createdById: userId,
-          reason: "Payment Rejection Failed - Order Not Found",
-          systemText: `Attempted to reject payment for non-existent order ${orderId}`,
-          userText: "Order not found."
-        })
-      );
-      return {
-        success: false,
-        message: "Order not found."
-      };
-    }
-    
-    // Find the payment to reject
-    const existingPayment = await prisma.payment.findUnique({
-      where: { id: paymentId }
-    });
-
-    if (!existingPayment) {
-      serverSideEffect(
-        () => createLog({
-          userId,
-          createdById: userId,
-          reason: "Payment Rejection Failed - Payment Not Found",
-          systemText: `Attempted to reject non-existent payment ${paymentId}`,
-          userText: "Payment not found."
-        })
-      );
-      return {
-        success: false,
-        message: "Payment not found."
-      };
-    }
-
-    // Update the existing payment record with rejected status
-    await prisma.payment.update({
-      where: { id: paymentId },
-      data: {
-        paymentStatus: PaymentStatus.DECLINED,
-        processedById: userId,
-        memo: existingPayment.memo 
-          ? `${existingPayment.memo} | Payment rejected: ${rejectionReason}` 
-          : `Payment rejected: ${rejectionReason}`
-      }
-    });
-
-    serverSideEffect(
-      () => createLog({
-        userId: order.customerId,
-        createdById: userId,
-        reason: "Payment Rejected",
-        systemText: `Rejected payment for order ${orderId}. Payment ID: ${paymentId}. Reason: ${rejectionReason}`,
-        userText: `Payment has been rejected. Reason: ${rejectionReason}`
-      }),
-      () => sendPaymentStatusEmail({
-        orderNumber: orderId,
-        customerName: `${order.customer.firstName} ${order.customer.lastName}`,
-        customerEmail: order.customer.email,
-        amount: Number(existingPayment.amount),
-        status: 'declined'
-      })
-    );
-
-    return {
-      success: true
-    };
-  } catch (error) {
-    serverSideEffect(
-      () => createLog({
-        userId,
-        createdById: userId,
-        reason: "Payment Rejection Error",
-        systemText: `Error rejecting payment for order ${orderId}: ${error instanceof Error ? error.message : 'Unknown error'}`,
-        userText: "An error occurred while rejecting the payment."
-      })
-    );
-
-    return {
-      success: false,
-      message: error instanceof Error ? error.message : "Failed to reject payment"
-    };
-  }
 }
