+++ src/app/admin/users/[email]/page.tsx
@@ -3,12 +3,7 @@ import { Metadata } from "next";
 import { FaShieldAlt } from "react-icons/fa";
 import { FaBagShopping, FaCreditCard  } from "react-icons/fa6";
 import { RiFilePaper2Line } from "react-icons/ri";
-import { UserProfileDetails } from "../components/user-profile-details";
-import { UserActions } from "../components/user-actions";
-import { UserPermissions } from "../components/user-permissions";
-import { UserActivityLogs } from "../components/user-activity-logs";
-import { UserOrderHistory } from "../components/user-order-history";
-import { UserPaymentHistory } from "../components/user-payment-history";
+import { UserProfileDetails , UserActions , UserPermissions , UserActivityLogs , UserOrderHistory , UserPaymentHistory } from "@/features/admin/users/components";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import PageAnimation from '@/components/public/page-animation';
 import PageTitle from '@/components/private/page-title';
