+++ src/features/admin/payments/actions/index.ts
@@ -1,0 +1,13 @@
+export * from './processPayment';
+export * from './refundPayment';
+export * from './validatePayment';
+export * from './rejectPayment';
+export * from '../payments.schema';
+
+// Export getters from payments.actions.ts
+export { 
+  getPayments,
+  getPaymentById,
+  getPaymentsByUser,
+  getPaymentsByOrderId
+} from '@/actions/payments.actions';
