+++ src/features/admin/payments/components/index.ts
@@ -1,0 +1,6 @@
+export { OffsitePayment } from './offsite-payment';
+export { OrdersPaymentTable } from './orders-payment-table';
+export { PaymentsContent } from './payments-content';
+export { TransactionHistory } from './transaction-history';
+export { OnsitePayment } from './onsite-payment';
+export { OrderPaymentModal, OrderPaymentModalWithQueryParams } from './order-payment-modal';
