+++ src/features/admin/payments/components/offsite-payment.tsx
@@ -1,5 +1,6 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
+'use client';
 
 import React, { useState, useEffect } from "react";
 import { useForm } from "react-hook-form";
