+++ src/features/admin/tickets/actions/index.ts
@@ -1,0 +1,7 @@
+export * from './createTicket';
+export * from './updateTicket';
+export * from './addTicketUpdate';
+export * from '../tickets.schema';
+
+// Export getters from ticket.actions
+export { getTickets, getTicketById } from '@/actions/ticket.actions';
