+++ src/features/admin/tickets/components/index.ts
@@ -1,0 +1,3 @@
+export { TicketsContent } from './tickets-content';
+export { TicketDialog } from './ticket-dialog';
+export { TicketList } from './ticket-list';
