+++ src/features/admin/users/actions/index.ts
@@ -1,0 +1,13 @@
+export * from './updateUserPermissions';
+export * from './resetPasswordForUser';
+export * from './updateUserDetails';
+export * from './assignManager';
+export * from '../users.schema';
+
+// Export getters from users.action.ts
+export { 
+  getClerkUserPublicData,
+  getClerkUserImageUrl,
+  getUsers,
+  getUser
+} from '@/actions/users.action';
